interface_test/requirements.txt
pymysql
requests
configparser
lxml
docx-mailmerge